Revert "Move handling to Interface class"
This reverts commit 09bdeec2362f773af5e8e2bf11e2c6703d0b693b.
